[analyzer] Dot shorthands: Update and add tests for const dot shorthands.
Updated the AST binary reader/writer to handle dot shorthand invocations otherwise we crash when we try to evaluate a method invocation. Added a test for that in `const_test.dart`.

Added a plethora of tests in `evaluation_test.dart` which led me to find a bug with not carrying over the `isDotShorthand` flag when we convert from a method invocation to a constructor invocation.

Unit tests passing.

Bug: https://github.com/dart-lang/sdk/issues/59835
Change-Id: Ide1cb7cffc0285ed7d54e9f7bc641d0d28c1d2e2
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/437881
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Kallen Tu <kallentu@google.com>