[CFE] Fix instrumenter so it works for aot
For a while it's been supported to do
`dart compile aot-snapshot <dillfile>`
meaning we can make the instrumentation thing run in aot too.

Except we couldn't because it crashed if you tried.

This CL sets up the dependency to the support library correctly (or at
least correctly enough) so that it now works.

Verified by doing

```
$ out/ReleaseX64/dart pkg/front_end/tool/flame/instrumenter.dart pkg/front_end/tool/compile.dart --count
$ out/ReleaseX64/dart-sdk/bin/dart compile aot-snapshot pkg/front_end/tool/compile.dart.dill.instrumented.dill
$ out/ReleaseX64/dart-sdk/bin/dartaotruntime pkg/front_end/tool/compile.dart.dill.instrumented.aot pkg/front_end/tool/compile.dart
```

Change-Id: I1d9b526b6ba2fed1ef81599fc9399041cc8b6522
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/439520
Commit-Queue: Jens Johansen <jensj@google.com>
Reviewed-by: Chloe Stefantsova <cstefantsova@google.com>