[analyzer] Remove format method; change callers to use formatList.
This method wasn't exposed through the analyzer public API, and wasn't
strictly necessary; all it did was wrap up all its arguments (except
the first) in a list and call `formatList`.

In the long term I intend to replace `formatList` with a richer and
more type-safe API that allows clients to supply message parameters
through named parameters. Removing the `format` method now will make
it easier to evolve towards that richer API.

Note that this change is not strictly a no-op; one of the call sites
(in
pkg/analysis_server/lib/src/services/refactoring/legacy/inline_local.dart)
was incorrectly wrapping its argument in a list and then calling
`format`, causing the argument to be wrapped in a second list. That
was clearly a mistake, so I went ahead and fixed it as part of this
CL.

Change-Id: I0796804ab949e66a129bfeec6dfe688dfa113eb2
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/440548
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>