Elements. Remove HasSinceSdkVersion, move 'sinceSdkVersion' into Element.
This simplifies both API and implementation; and unblocks another
simplification with removing `TypeDefiningElementImpl`.

There are no uses of `HasSinceSdkVersion` interface, and even
`sinceSdkVersion` at all in google3. So, we might consider removing it
at all. But I imagine that it might be theoretically useful, so maybe
not. OTOH, given that its use it so spare, I think that it does not
deserve the complexity.

To be landed after 7.7.0 from
https://dart-review.googlesource.com/c/sdk/+/440683 is published.

Change-Id: I69bb2d5aeba8a463b1fb30704f5877527f46e5e9
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/440600
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Paul Berry <paulberry@google.com>