[analyzer] Remove test class extending DiagnosticReporter.
Previous to this commit, `lint_rule_test.dart` contained a class,
`CollectingReporter`, which extended `DiagnosticReporter` and modified
its behavior so that instead of reporting the error to its listener,
it simply recorded the error code, so that tests could verify that an
error with the appropriate error code was reported.

It's not really necessary to have a whole class for this; the same
effect can be more easily accomplished by simply looking at the error
recorded by the `GatheringDiagnosticListener`, and extracting the
error code from it.

Eliminating the `CollectingReporter` class should make it easier to
refactor and add functionality to the `DiagnosticReporter` class,
which I intend to do as part of re-working the analyzer's error
reporting mechanism to be more type safe.

Change-Id: Ife8172da39bc9157a917c9b2d36cb3ee53dde47b
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/441020
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>
Auto-Submit: Paul Berry <paulberry@google.com>