Add tests for most `displayString()` methods.
Significantly improve test coverage of `display_string_builder.dart`.

Fix an error involving a missing space when reading the enclosing
fragment of a local variable declaration.

Change the formatting of `ConstructorElement` to print only the type and
not the name to avoid duplication, as they were the same.

Change the formatting of `TypeAliasElement` to just print the name of
the aliased type, rather than the whole aliased element - as this rarely
provided useful information.

Change-Id: Ie1e4a3649063c2b3662ef7bd21327c17a5b1456b
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/441082
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Paul Berry <paulberry@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>