Remove `extract_errors_to_yaml.dart`.
This was a temporary tool that I used in the process of creating the
analyzer's `messages.yaml` file. I intended to delete it once the
`messages.yaml` file was complete, but forgot.

Deleting the temporary tool will make it easier to make improvements
to the `messages.yaml` format.

Change-Id: I9ec638fbf8a6236ff9b4d8cea6aadca913768d53
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/441680
Auto-Submit: Paul Berry <paulberry@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>