[analyzer] Change ErrorCodeInfo.toAnalyzerCode to accept ErrorClassInfo.
Previously, `ErrorCodeInfo.toAnalyzerCode` didn't know anything about
the error class it was generating an analyzer code for, other than the
class name (which was passed in via a `className` parameter). Changing
this parameter to an `ErrorClassInfo` type paves the way for a
follow-up CL in which I plan to generalize
`ErrorCodeInfo.toAnalyzerCode` so that it can generate analyzer error
codes that support `.withArguments()` (similar to how CFE error
formatting works).

To make this work, I had to create an instance of `ErrorClassInfo` in
`generate_lints.dart`, to represent the linter class `LinterLintCode`.

Change-Id: I160e27a46602da13f4757393fdaad4bc2f15d50e
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/442165
Auto-Submit: Paul Berry <paulberry@google.com>
Commit-Queue: Samuel Rawlins <srawlins@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>