Elements. Write nameOffset, firstTokenOffset, offset for fragments.
With changes to the way that we set `firstTokenOffset` that I have in
mind, I think it would be useful to write actual values.

After reviewing the current code for writing "best offset", I also
think that it is better just write the value, and not try to be
smart about it.

And for consistency we use `(nameOffset:6)` for the name offset.

Change-Id: I83772ea7497d5edf2b6c5ae6448833956be954fc
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/442006
Reviewed-by: Paul Berry <paulberry@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>