[VM/Service] Tweak profile_startup_cli_flag_test to make it less flaky
The `--profile-period` argument is used when calculating a sample buffer
size that satisfies the `--sample-buffer-duration` argument, so this CL
changes the `--profile-period` argument we pass from 100 to 500, because
the profiler is more likely to actually hit that rate of sampling.

This CL also changes `testeeMain` to spin for 10 seconds instead of 5
seconds, to make it more likely to fill up the sample buffer.

Issue: https://github.com/dart-lang/sdk/issues/61191
Change-Id: Ic18b47f86bd2495b7df1af4d7f886fb11aa42812
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/442242
Commit-Queue: Derek Xu <derekx@google.com>
Reviewed-by: Ben Konyi <bkonyi@google.com>