Adjust unrelated_type_equality_checks on mixins and extension types
The lint unrelated_type_equality_checks previously considered all mixins
and extension types to be (possibly) related (even any mixin and any
extension type, not just any pair of mixins and any pair of extension
types). This would not fit the expectations for this lint, it should
certainly consider a mixin as unrelated to an extension type that isn't
a subtype of the mixin, and similarly for any pair of mixins that aren't
subtype related, and any pair of extension types that aren't subtype
related.

This PR changes the lint (that is, the underlying extension method
`interfaceTypesAreUnrelated`) such that these pairs are considered
unrelated.

Bug: #59423
Change-Id: I370e98eb17c60ab50f1118b6e841d7eee7291e54
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/440640
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Samuel Rawlins <srawlins@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>