[cfe] Compute name spaces for dill directly
This changes the computation of the name space for dill libraries and
declarations to directly filling out the underlying maps.

This is a step towards avoiding computation of lookup results for
conflicting upon lookup. Name spaces from dill can have no conflicts
so the computation is not needed.

The remaining piece is name spaces for exports and prefixes. These
also cannot have conflicts but instead of ambiguous lookups which only
trigger errors on use.

Change-Id: Ibe93620c66f4820027eb7b3e71fafab96c6b5ba6
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/441940
Reviewed-by: Chloe Stefantsova <cstefantsova@google.com>
Commit-Queue: Johnni Winther <johnniwinther@google.com>