Elements. Convert internal Element interfaces, known as '2OrMember', into mixins.
It was alreay a mixin for formal parameters, and I cannot think a
disadvantage to make all other mixins. There is a possible advantage
that we might provide also some methods with implementations, although
currently we do not.

This makes them consistent, which is nice IMHO.

Change-Id: Ic8382c68b7e2719f94aeb01ff6553f6ff5249c70
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/442506
Reviewed-by: Paul Berry <paulberry@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>