[cfe] Remove ProblemBuilder and clean up lookup methods
This removes the ProblemBuilder which is no longer needed. With the removal, the `lookupLocal` and `lookupLocalMember` are no longer different and are merged to `lookup`. Furthermore since the `fileUri` is not used and `fileOffset` is only used on `LocalScope`, the signature of `lookup` is reduced to only taking the name, and `lookup` on `LocalScope` takes an optional `fileOffset`.

Change-Id: Ib81e6069dc078c62af0288349a31c36c962f41ae
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/442681
Reviewed-by: Chloe Stefantsova <cstefantsova@google.com>
Commit-Queue: Johnni Winther <johnniwinther@google.com>