[analyzer] Dot shorthands: Issue 61219 - Recovery for selector chains in LHS of assignment.
This CL adds caching of a context type in the particular case that we have a dot shorthand in a property access in the LHS of an assignment.

This scenario would previously crash upon seeing that we have no cached context type. It's still not valid code, but at least we're not crashing when we analyze it.

Bug: https://github.com/dart-lang/sdk/issues/61219
Change-Id: I0159b90de7a53a8a1120e0ba2c7be5d7758cc17d
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/442803
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Kallen Tu <kallentu@google.com>