Fine. Support for all top-level elements in getElementId().
This adds previously missed classes, which was not noticed previously
because the implementation was always incomplete, and I wanted to have
something runnable. I started running fine-grained dependencies over
Flutter and uncovered a few holes. So, now it is time to add stricter
checks and complete the list of top-level elements.

Change-Id: If5e5d561817cd8e61fb7e666338355cbeca35974
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/443620
Reviewed-by: Paul Berry <paulberry@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>