Fine. Support for more constant expressions.
Found a few missing cases while running over Flutter.

We only need to support valid constant expressions, because we serialize
only constant expressions. So, it is OK to say that if there was invalid
expression, it does not match with any expression, we will consider it
invalidated until the user fixes the code. We support this with
`isValid` flag now.

* Added support for `identical()`.
* Added support for `ImportPrefixReference`.
* Added support for `SymbolLiteral`.

Change-Id: Ief8714f281cd49b62d242bf85e92c2531b79af95
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/443623
Reviewed-by: Johnni Winther <johnniwinther@google.com>