[CFE] Tests for the coverage tests
More specifically, tests for the coverage merger tool, i.e. allowing us
to write tests for missing coverage being reported in a certain way, and
for the automatic update of comments working like we want it to.

As this testing is added after-the-fact the (initial at least) testing
is very sparse.

Change-Id: I5ae372b4ad56511859ad6f616c99835d1ccd140f
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/443960
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Jens Johansen <jensj@google.com>